pyproject.toml
[tool.black]
line-length = 120
target_version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | versions
  )/
)
'''


[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true


[tool.poetry]
name = "asana_bot"
version = "0.1.0"
description = ""
authors = ["bekzod-fayzikuloff <fayzikuloff.b@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asana = "^0.10.13"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.37"
schedule = "^1.1.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
mypy = "^0.960"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
